reorganized axios related code from main.js + fix token bugs
