+++ toktik-frontend/src/views/AboutView.vue
@@ -15,6 +15,7 @@
 
 <script>
 import router from '@/router';
+import axios from '@/axios';
 import VideoPlayer from '@/components/VideoPlayer.vue';
 
 export default {
@@ -28,7 +29,6 @@ export default {
   },
   setup() {
     const token = localStorage.getItem("jwt-token")
-    console.log(token)
     if (token) {
       return { token: token }
     }
@@ -41,7 +41,7 @@ export default {
       const token_refresh = localStorage.getItem( 'jwt-token-refresh' );
       if (token) {
         localStorage.removeItem('jwt-token')
-        this.axios.defaults.headers.common['Authorization'] = 'Bearer ';
+        axios.defaults.headers.common['Authorization'] = '';
       }
       if (token_refresh) {
         localStorage.removeItem('jwt-token-refresh')
