+++ toktik-frontend/src/views/RegisterView.vue
@@ -82,6 +82,7 @@
 
 <script>
 import router from '@/router';
+import axios from '@/axios';
 
 export default {
   data() {
@@ -102,7 +103,7 @@ export default {
         formData.append("email", this.email);
         formData.append("password", this.password);
         
-        const response = await this.axios({
+        const response = await axios({
           method: "POST",
           url: `/api/auth/register/`,
           data: formData,
